Changed the structure of data transfer and created an underlying data structure for telemetry collection

+++ pod-src/include/mcp23017.h
@@ -59,6 +59,6 @@ class Iox {
 /*int ioxSetDir(IoxPin pin, IoxDir dir);*/
 /*int ioxRead(IoxPin pin);*/
 /*int ioxWrite(IoxPin pin, IoxVal val);*/
-float testWriteIOX();
-float testReadIOX();
+uint16_t testWriteIOX();
+uint16_t testReadIOX();
 #endif
