+++ pod-src/src/mcp23017.cpp
@@ -77,7 +77,7 @@ int Iox::write(IoxPin pin, IoxVal val) {
 
 
 /* TESTS */
-float testWriteIOX() {
+uint16_t testWriteIOX() {
     I2C i2c(PB_7, PB_6);
     Iox iox(&i2c, 0x20);
     for (int i = 0; i < 16; i++) {
@@ -86,10 +86,10 @@ float testWriteIOX() {
         if (iox.write((IoxPin)i, HIGH)) 
             pc.printf("ERROR WRITING PIN\n\r");
     }
-    return 0.0;
+    return 0;
 }
 
-float testReadIOX() {
+uint16_t testReadIOX() {
     I2C i2c(PB_7, PB_6);
     Iox iox(&i2c, 0x20);
     for (int i = 0; i < 16; i++) {
@@ -97,6 +97,6 @@ float testReadIOX() {
             pc.printf("ERROR SETTING DIR\n\r");
         pc.printf("READ FROM PIN: %d\n\r", (int)iox.read((IoxPin)i));
     }
-    return 0.0;
+    return 0;
 
 }
